rename: 함수 및 파일명 내 room을 box로 변경
